Also added an interface for AmazonTitleIndexStore
Refactored class names
