--- src/main/java/org/atlasapi/remotesite/amazon/AmazonPersistenceModuleInterface.java
@@ -1,8 +1,0 @@
-package org.atlasapi.remotesite.amazon;
-
-import org.atlasapi.remotesite.amazon.indexer.AmazonTitleIndexStore;
-
-public interface AmazonPersistenceModuleInterface {
-
-    AmazonTitleIndexStore amazonTitleIndexStore();
-}
