+++ src/test/java/org/atlasapi/remotesite/amazon/AmazonTitleIndexStoreImplTest.java
@@ -4,28 +4,28 @@ import com.google.common.collect.ImmutableSet;
 import com.metabroadcast.common.persistence.MongoTestHelper;
 import com.metabroadcast.common.persistence.mongo.DatabasedMongo;
 import org.atlasapi.remotesite.amazon.indexer.AmazonTitleIndexEntry;
-import org.atlasapi.remotesite.amazon.indexer.AmazonTitleIndexStore;
+import org.atlasapi.remotesite.amazon.indexer.AmazonTitleIndexStoreImpl;
 import org.junit.Before;
 import org.junit.Test;
 
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertThat;
 
-public class AmazonTitleIndexStoreTest {
+public class AmazonTitleIndexStoreImplTest {
 
-    private AmazonTitleIndexStore amazonTitleIndexStore;
+    private AmazonTitleIndexStoreImpl amazonTitleIndexStoreImpl;
 
     @Before
     public void setUp() {
         DatabasedMongo mongo = MongoTestHelper.anEmptyTestDatabase();
-        amazonTitleIndexStore = new AmazonTitleIndexStore(mongo);
+        amazonTitleIndexStoreImpl = new AmazonTitleIndexStoreImpl(mongo);
     }
 
     @Test
     public void testRetrievingTitleIndexEntryIsEqual() {
         AmazonTitleIndexEntry entry = new AmazonTitleIndexEntry("title", ImmutableSet.of("uri1", "uri2"));
-        amazonTitleIndexStore.createOrUpdateIndex(entry);
-        AmazonTitleIndexEntry retrievedEntry = amazonTitleIndexStore.getIndexEntry(entry.getTitle());
+        amazonTitleIndexStoreImpl.createOrUpdateIndex(entry);
+        AmazonTitleIndexEntry retrievedEntry = amazonTitleIndexStoreImpl.getIndexEntry(entry.getTitle());
         assertThat(retrievedEntry, is(entry));
     }
